restaurants list apply mock up theme

+++ app/src/main/java/com/cleanup/go4lunch/ui/list/PlacesListViewState.kt
@@ -11,6 +11,6 @@ data class PlacesListViewState(
     val colleagues: String,
     val image: Bitmap?,
     val hours: String,
-    val likes: Int
+    val likes: Float
 )
